refactor(accounts): split the accounts composable by verb
`useBlockchainAccounts` had grown to 323 lines and 11 functions across four
unrelated concerns, and hit the 20-import cap twice while the activity work
was going on. Its four consumers each destructured exactly one of those
concerns, so the god object bought nothing.

Split into `use-account-additions`, `use-account-edits`,
`use-account-removals` and `use-account-fetching`, each importing only what
it needs. No facade: with a 1:1 consumer mapping an extra indirection would
just re-inherit every import.

Two duplications collapsed on the way: three identical removal-failure
notifications became one helper, and the two private fetch helpers stopped
returning values nothing read.

Two spec mocks turned out to be dead — `modules/core/messaging` never
imported this module, and its mock named functions the composable has not
had for some time.
