+++ frontend/app/src/modules/accounts/blockchain/use-account-delete.ts
@@ -5,7 +5,7 @@ import type {
 } from '@/modules/accounts/blockchain-accounts';
 import { Blockchain } from '@rotki/common';
 import { getAccountAddress, isXpubAccount } from '@/modules/accounts/account-utils';
-import { useBlockchainAccounts } from '@/modules/accounts/use-blockchain-accounts';
+import { useAccountRemovals } from '@/modules/accounts/use-account-removals';
 import { useBlockchainAccountsStore } from '@/modules/accounts/use-blockchain-accounts-store';
 import { useEthStaking } from '@/modules/accounts/use-eth-staking';
 import { useBalancesStore } from '@/modules/balances/use-balances-store';
@@ -129,7 +129,7 @@ export function useAccountDelete(): UseAccountDeleteReturn {
   const { accounts } = storeToRefs(useBlockchainAccountsStore());
   const { balances } = storeToRefs(useBalancesStore());
   const { deleteEth2Validators } = useEthStaking();
-  const { deleteXpub, removeAccount, removeAgnosticAccount } = useBlockchainAccounts();
+  const { deleteXpub, removeAccount, removeAgnosticAccount } = useAccountRemovals();
   const { t } = useI18n({ useScope: 'global' });
   const { show } = useConfirmStore();
   const { getChainName } = useSupportedChains();
